fix: 404 styling
